+++ src/metrics/weighted-size.d.ts
@@ -1,6 +1,9 @@
-import Graph from 'graphology-types';
+import Graph, {Attributes, EdgeMapper} from 'graphology-types';
 
-export default function weightedSize(
-  graph: Graph,
-  weightAttribute?: string
+export default function weightedSize<
+  NodeAttributes extends Attributes = Attributes,
+  EdgeAttributes extends Attributes = Attributes
+>(
+  graph: Graph<NodeAttributes, EdgeAttributes>,
+  getEdgeWeight?: string | EdgeMapper<NodeAttributes, EdgeAttributes>
 ): number;